% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{DecryptFile}
\alias{DecryptFile}
\title{Decrypt a file with the DES algorithm}
\usage{
DecryptFile(sourceFile, decryptedFile, key, SUNOS = FALSE, HEX_KEY = FALSE,
  ECB = FALSE, UUENC = FALSE)
}
\description{
Decrypt a file with the DES algorithm
}

