% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{DecryptFileTripleDES}
\alias{DecryptFileTripleDES}
\title{Decrypt a file with the Triple DES algorithm}
\usage{
DecryptFileTripleDES(sourceFile, decryptedFile, key, ECB = FALSE,
  UUENC = FALSE)
}
\description{
Decrypt a file with the Triple DES algorithm
}

