% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{EncryptFile}
\alias{EncryptFile}
\title{Encrypt a file with the DES algorithm}
\usage{
EncryptFile(sourceFile, encryptedFile, key, SUNOS = FALSE, HEX_KEY = FALSE,
  ECB = FALSE, UUENC = FALSE)
}
\description{
Encrypt a file with the DES algorithm
}

