% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{EncryptFileTripleDES}
\alias{EncryptFileTripleDES}
\title{Encrypt a file with the Triple DES algorithm}
\usage{
EncryptFileTripleDES(sourceFile, encryptedFile, key, ECB = FALSE,
  UUENC = FALSE)
}
\description{
Encrypt a file with the Triple DES algorithm
}

