% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{EncryptFileTripleDES}
\alias{EncryptFileTripleDES}
\title{Encrypt a file with the Triple DES algorithm}
\usage{
EncryptFileTripleDES(sourceFile, encryptedFile, key, ECB = FALSE,
  UUENC = FALSE)
}
\arguments{
\item{sourceFile}{The path the the file that should be encrypted}

\item{encryptedFile}{The path to the target file}

\item{key}{The 56bit HEX key}

\item{ECB}{Use ECB?}

\item{UUENC}{Whether the files should be UU encoded}
}
\description{
Encrypt a file with the Triple DES algorithm
}

